добавлена проверка на удаление самого себя

+++ server/modules/admin/controllers.ts
@@ -3,10 +3,15 @@ import mongoose from "mongoose";
 import { handleError } from "../../tools";
 import User from "../auth/models/User";
 
-interface IRequest extends express.Request {
+interface IEditUserRequest extends express.Request {
   body: {
     _id: string;
   }
+}
+
+interface IDeleteUserRequest extends IEditUserRequest {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  user?: any,
 }
 
 export const users = async (req: express.Request, res: express.Response): Promise<unknown> => {
@@ -18,7 +23,7 @@ export const users = async (req: express.Request, res: express.Response): Promis
   }
 };
 
-export const editUser = async (req: IRequest, res: express.Response): Promise<unknown> => {
+export const editUser = async (req: IEditUserRequest, res: express.Response): Promise<unknown> => {
   try {
     const _id = mongoose.Types.ObjectId(req.body._id);
     await User.updateOne({ _id }, req.body);
@@ -28,8 +33,11 @@ export const editUser = async (req: IRequest, res: express.Response): Promise<un
   }
 };
 
-export const deleteUser = async (req: IRequest, res: express.Response): Promise<unknown> => {
+export const deleteUser = async (req: IDeleteUserRequest, res: express.Response): Promise<unknown> => {
   try {
+    if (req.body._id === req.user._id + "") {
+      return res.status(400).end("Вы не можете удались самого себя!");
+    }
     const _id = mongoose.Types.ObjectId(req.body._id);
     await User.deleteOne({ _id });
     return res.status(200).end("Пользователь успешно удален.");
